+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 1.0.3 under development
 --------------------------
 
+- Tính năng #10: Thêm validator `bankvn` hổ trợ kiểm tra mã ngân hàng (vuongxuongminh)
 - Tính năng #8: Thêm thuộc tính `autoDisableControllerCsrfValidation` tại `\yiiviet\payment\VerifyFilter` (vuongxuongminh). 
 - Lỗi #7: xác minh `checksum` tại phương thức `verifyPurchaseSuccess` của Bảo Kim (vuongxuongminh).
 - Lỗi #6: `checksum` vẫn bị yêu cầu đối với phương thức `purchase` pro của Bảo Kim (vuongxuongminh).